Added Backup Overview List

+++ BackupApp/src/main/java/org/secuso/privacyfriendlybackup/data/room/model/StoredBackupMetaData.kt
@@ -12,7 +12,7 @@ import java.util.*
 @Entity(indices = [Index(value = ["packageName"])])
 data class StoredBackupMetaData(
     @PrimaryKey(autoGenerate = true)
-    val _id : Int = 0,
+    val _id : Long = 0,
     val packageName : String,
     val timestamp : Date,
     val storageService : StorageType,
